Fix bug in debug node detection.
Examples can be indicated as the "debug" example, which can be set on a single example to make it the root example of the catalog app. This can speed up iteration cycles as you don't need to navigate to the example each time you want to test it.

There was a bug where this metadata was not being surfaced to the root of the tree however, because the metadata being inspected for debug status was not referring to the appropriate metadata object.

This CL fixes this bug by ensuring that we look at the node's actual metadata fields when determining whether a node should be deemed the "debug" node.

PiperOrigin-RevId: 563536785